+++ cmd/pam.go
@@ -16,22 +16,13 @@ package cmd
 
 import (
 	"context"
-	_ "embed"
 	"encoding/json"
 	"fmt"
-	"io"
-	"net/http"
-	"os"
-	"strconv"
-	"strings"
 
 	keyfactor "github.com/Keyfactor/keyfactor-go-client/v3/api"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
-
-//go:embed pam_types.json
-var EmbeddedPAMTypesJSON []byte
 
 type JSONImportableObject interface {
 	keyfactor.Provider |
@@ -49,321 +40,6 @@ var pamCmd = &cobra.Command{
 	Long: `Privileged Access Management (PAM) functionality in Keyfactor Web APIs allows for configuration of third 
 party PAM providers to secure certificate stores. The PAM component of the Keyfactor API includes methods necessary to 
 programmatically create, delete, edit, and list PAM Providers.`,
-}
-
-var pamTypesGetCmd = &cobra.Command{
-	Use:   "types-get",
-	Short: "Get a specific defined PAM Provider type by ID or Name.",
-	Long:  "Get a specific defined PAM Provider type by ID or Name.",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		cmd.SilenceUsage = true
-		isExperimental := false
-		// Specific flags
-		pamProviderTypeId, _ := cmd.Flags().GetString("id")
-		pamProviderTypeName, _ := cmd.Flags().GetString("name")
-		// Debug + expEnabled checks
-		informDebug(debugFlag)
-		debugErr := warnExperimentalFeature(expEnabled, isExperimental)
-		if debugErr != nil {
-			return debugErr
-		}
-		// Log flags
-		log.Info().Str("name", pamProviderTypeName).
-			Str("id", pamProviderTypeId).
-			Msg("get PAM Provider Type")
-		// Authenticate
-		kfClient, cErr := initClient(false)
-		if cErr != nil {
-			return cErr
-		}
-		if pamProviderTypeId == "" && pamProviderTypeName == "" {
-			cmd.Usage()
-			return fmt.Errorf("must supply either a PAM Provider Type `--id` or `--name` to get")
-		}
-
-		// CLI Logic
-		if pamProviderTypeId == "" && pamProviderTypeName != "" {
-			// Get ID from Name
-			log.Debug().Str("pamProviderTypeName", pamProviderTypeName).
-				Msg("call: GetPAMProviderTypeByName()")
-			pamProviderType, getErr := kfClient.GetPAMProviderTypeByName(pamProviderTypeName)
-			log.Debug().Msg("returned: GetPAMProviderTypeByName()")
-			if getErr != nil {
-				log.Error().Err(getErr).Send()
-				return getErr
-			}
-			if pamProviderType != nil {
-				output, mErr := json.Marshal(pamProviderType)
-				if mErr != nil {
-					log.Error().Err(mErr).Send()
-					return mErr
-				}
-				log.Info().Str("output", string(output)).
-					Msg("successfully retrieved PAM provider type")
-				outputResult(output, outputFormat)
-				return nil
-			}
-		}
-		pamProviderType, getErr := kfClient.GetPAMProviderType(pamProviderTypeId)
-		if getErr != nil {
-			log.Error().Err(getErr).Send()
-			return getErr
-		}
-		output, mErr := json.Marshal(pamProviderType)
-		if mErr != nil {
-			log.Error().Err(mErr).Send()
-			return mErr
-		}
-		log.Info().Str("output", string(output)).
-			Msg("successfully retrieved PAM provider type")
-		outputResult(output, outputFormat)
-		return nil
-	},
-}
-
-var pamTypesListCmd = &cobra.Command{
-	Use:   "types-list",
-	Short: "Returns a list of all available PAM provider types.",
-	Long:  "Returns a list of all available PAM provider types.",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		cmd.SilenceUsage = true
-		isExperimental := false
-
-		informDebug(debugFlag)
-		debugErr := warnExperimentalFeature(expEnabled, isExperimental)
-		if debugErr != nil {
-			return debugErr
-		}
-
-		// Log flags
-		log.Info().Msg("list PAM Provider Types")
-
-		// Authenticate
-		kfClient, clientErr := initClient(false)
-		if clientErr != nil {
-			return clientErr
-		}
-
-		//// CLI Logic
-		//log.Debug().Msg("call: PAMProviderGetPamProviderTypes()")
-		//pamTypes, httpResponse, err := sdkClient.PAMProviderApi.
-		//	PAMProviderGetPamProviderTypes(context.Background()).
-		//	XKeyfactorRequestedWith(XKeyfactorRequestedWith).
-		//	XKeyfactorApiVersion(XKeyfactorApiVersion).
-		//	Execute()
-		pamTypes, err := kfClient.ListPAMProviderTypes()
-		log.Debug().Msg("returned: PAMProviderGetPamProviderTypes()")
-		if err != nil {
-			log.Error().Err(err).
-				Msg("error listing PAM provider types")
-			return err
-		}
-
-		log.Debug().Msg("Converting PAM Provider Types response to JSON")
-		jsonString, mErr := json.Marshal(pamTypes)
-		if mErr != nil {
-			log.Error().Err(mErr).Send()
-			return mErr
-		}
-		log.Info().
-			Msg("successfully listed PAM provider types")
-		outputResult(jsonString, outputFormat)
-		return nil
-	},
-}
-
-var pamTypesCreateCmd = &cobra.Command{
-	Use:   "types-create",
-	Short: "Creates a new PAM provider type.",
-	Long: `Creates a new PAM Provider type, currently only supported from JSON file and from GitHub. To install from 
-Github. To install from GitHub, use the --repo flag to specify the GitHub repository and optionally the branch to use. 
-NOTE: the file from Github must be named integration-manifest.json and must use the same schema as 
-https://github.com/Keyfactor/hashicorp-vault-pam/blob/main/integration-manifest.json. To install from a local file, use
---from-file to specify the path to the JSON file.`,
-	RunE: func(cmd *cobra.Command, args []string) error {
-		cmd.SilenceUsage = true
-		isExperimental := false
-
-		// Specific flags
-		pamConfigFile, _ := cmd.Flags().GetString(FlagFromFile)
-		pamProviderName, _ := cmd.Flags().GetString("name")
-		repoName, _ := cmd.Flags().GetString("repo")
-		branchName, _ := cmd.Flags().GetString("branch")
-
-		// Debug + expEnabled checks
-		informDebug(debugFlag)
-		debugErr := warnExperimentalFeature(expEnabled, isExperimental)
-		if debugErr != nil {
-			return debugErr
-		}
-
-		// Log flags
-		log.Info().Str("name", pamProviderName).
-			Str("repo", repoName).
-			Str("branch", branchName).
-			Msg("create PAM Provider Type")
-
-		// Authenticate
-		kfClient, cErr := initClient(false)
-		//sdkClient, cErr := initGenClient(false)
-		if cErr != nil {
-			return cErr
-		}
-
-		// Check required flags
-		if pamConfigFile == "" && repoName == "" {
-			cmd.Usage()
-			return fmt.Errorf("must supply either a config `--from-file` or a `--repo` GitHub repository to get file from")
-		} else if pamConfigFile != "" && repoName != "" {
-			cmd.Usage()
-			return fmt.Errorf("must supply either a config `--from-file` or a `--repo` GitHub repository to get file from, not both")
-		}
-
-		// CLI Logic
-
-		var pamProviderType *keyfactor.ProviderTypeCreateRequest
-		var err error
-		if repoName != "" {
-			// get JSON config from integration-manifest on GitHub
-			log.Debug().
-				Str("pamProviderName", pamProviderName).
-				Str("repoName", repoName).
-				Str("branchName", branchName).
-				Msg("call: GetTypeFromInternet()")
-			pamProviderType, err = GetTypeFromInternet(pamProviderName, repoName, branchName, pamProviderType)
-			log.Debug().Msg("returned: GetTypeFromInternet()")
-			if err != nil {
-				log.Error().Err(err).Send()
-				return err
-			}
-		} else {
-			log.Debug().Str("pamConfigFile", pamConfigFile).
-				Msg(fmt.Sprintf("call: %s", "GetTypeFromConfigFile()"))
-			pamProviderType, err = GetTypeFromConfigFile(pamConfigFile, pamProviderType)
-			log.Debug().Msg(fmt.Sprintf("returned: %s", "GetTypeFromConfigFile()"))
-			if err != nil {
-				log.Error().Err(err).Send()
-				return err
-			}
-		}
-
-		if pamProviderName != "" {
-			pamProviderType.Name = pamProviderName
-		}
-
-		log.Info().Str("pamProviderName", pamProviderType.Name).
-			Msg("creating PAM provider type")
-
-		log.Debug().Msg("call: PAMProviderCreatePamProviderType()")
-		createdPamProviderType, rErr := kfClient.CreatePAMProviderType(pamProviderType)
-		log.Debug().Msg("returned: PAMProviderCreatePamProviderType()")
-		if rErr != nil {
-			log.Error().Err(rErr).Send()
-			return rErr
-		}
-
-		log.Debug().Msg("Converting PAM Provider Type response to JSON")
-		jsonString, mErr := json.Marshal(createdPamProviderType)
-		if mErr != nil {
-			log.Error().Err(mErr).Send()
-			return mErr
-		}
-		log.Info().Str("output", string(jsonString)).
-			Msg("successfully created PAM provider type")
-		outputResult(jsonString, outputFormat)
-		return nil
-	},
-}
-
-var pamTypesDeleteCmd = &cobra.Command{
-	Use:   "types-delete",
-	Short: "Deletes a defined PAM Provider type by ID or Name.",
-	Long:  "Deletes a defined PAM Provider type by ID or Name.",
-	RunE: func(cmd *cobra.Command, args []string) error {
-		cmd.SilenceUsage = true
-		isExperimental := false
-		// Specific flags
-		pamProviderTypeId, _ := cmd.Flags().GetString("id")
-		pamProviderTypeName, _ := cmd.Flags().GetString("name")
-		deleteAll, _ := cmd.Flags().GetBool("all")
-		// Debug + expEnabled checks
-		informDebug(debugFlag)
-		debugErr := warnExperimentalFeature(expEnabled, isExperimental)
-		if debugErr != nil {
-			return debugErr
-
-		}
-		// Log flags
-		log.Info().Str("name", pamProviderTypeName).
-			Str("id", pamProviderTypeId).
-			Msg("delete PAM Provider Type")
-		// Authenticate
-		kfClient, cErr := initClient(false)
-		if cErr != nil {
-			return cErr
-		}
-		// CLI Logic
-
-		if deleteAll {
-			log.Info().Msg("deleting ALL PAM Provider Types")
-			pamProviderTypes, listErr := kfClient.ListPAMProviderTypes()
-			if listErr != nil {
-				log.Error().Err(listErr).Send()
-				return listErr
-			}
-			if pamProviderTypes == nil || len(*pamProviderTypes) == 0 {
-				log.Info().Msg("no PAM provider types to delete")
-				outputResult("No PAM provider types to delete", outputFormat)
-			}
-			for _, pamProviderType := range *pamProviderTypes {
-				log.Debug().Str("pamProviderTypeId", pamProviderType.Id).
-					Msg("call: PAMProviderDeletePamProviderType()")
-				delErr := kfClient.DeletePAMProviderType(pamProviderType.Id)
-				if delErr != nil {
-					log.Error().Err(delErr).Send()
-					outputError(delErr, false, outputFormat)
-					continue
-				}
-				log.Info().Str("id", pamProviderType.Id).Str("name", *pamProviderType.Name).
-					Msg("successfully deleted PAM provider type")
-				outputResult(fmt.Sprintf("Deleted PAM provider type with ID %s", pamProviderType.Id), outputFormat)
-			}
-			log.Info().Msg("successfully deleted ALL PAM provider types")
-			outputResult(fmt.Sprintf("Deleted ALL %d PAM provider types", len(*pamProviderTypes)), outputFormat)
-			return nil
-		}
-		if pamProviderTypeId == "" && pamProviderTypeName == "" {
-			cmd.Usage()
-			return fmt.Errorf("must supply either a PAM Provider Type `--id` or `--name` to delete")
-		}
-
-		if pamProviderTypeId == "" && pamProviderTypeName != "" {
-			// Get ID from Name
-			log.Debug().Str("pamProviderTypeName", pamProviderTypeName).
-				Msg("call: GetPAMProviderTypeByName()")
-			pamProviderType, getErr := kfClient.GetPAMProviderTypeByName(pamProviderTypeName)
-			log.Debug().Msg("returned: GetPAMProviderTypeByName()")
-			if getErr != nil {
-				log.Error().Err(getErr).Send()
-				return getErr
-			}
-			pamProviderTypeId = pamProviderType.Id
-		}
-
-		log.Debug().Str("pamProviderTypeId", pamProviderTypeId).
-			Msg("call: PAMProviderDeletePamProviderType()")
-		delErr := kfClient.DeletePAMProviderType(pamProviderTypeId)
-		if delErr != nil {
-			log.Error().Err(delErr).Send()
-			return delErr
-		}
-
-		log.Info().Str("name", pamProviderTypeName).
-			Str("id", pamProviderTypeId).
-			Msg("successfully deleted PAM provider type")
-		outputResult(fmt.Sprintf("Deleted PAM provider type with ID %s", pamProviderTypeId), outputFormat)
-		return nil
-	},
 }
 
 var pamProvidersListCmd = &cobra.Command{
@@ -474,30 +150,6 @@ var pamProvidersGetCmd = &cobra.Command{
 		outputResult(jsonString, outputFormat)
 		return nil
 	},
-}
-
-func checkBug63171(cmdResp *http.Response, operation string) error {
-	if cmdResp != nil && cmdResp.StatusCode == 200 {
-		defer cmdResp.Body.Close()
-		// .\Admin
-		productVersion := cmdResp.Header.Get("X-Keyfactor-Product-Version")
-		log.Debug().Str("productVersion", productVersion).Msg("Keyfactor Command Version")
-		majorVersionStr := strings.Split(productVersion, ".")[0]
-		// Try to convert to int
-		majorVersion, err := strconv.Atoi(majorVersionStr)
-		if err == nil && majorVersion >= 12 {
-			// TODO: Pending resolution of this bug: https://dev.azure.com/Keyfactor/Engineering/_workitems/edit/63171
-			errMsg := fmt.Sprintf(
-				"PAM Provider %s is not supported in Keyfactor Command version 12 and later, "+
-					"please use the Keyfactor Command UI to create PAM Providers", operation,
-			)
-			oErr := fmt.Errorf(errMsg)
-			log.Error().Err(oErr).Send()
-			outputError(oErr, true, outputFormat)
-			return oErr
-		}
-	}
-	return nil
 }
 
 var pamProvidersCreateCmd = &cobra.Command{
@@ -706,190 +358,6 @@ var pamProvidersDeleteCmd = &cobra.Command{
 	},
 }
 
-func GetPAMTypeInternet(providerName string, repo string, branch string) (interface{}, error) {
-	log.Debug().Str("providerName", providerName).
-		Str("repo", repo).
-		Str("branch", branch).
-		Msg("entered: GetPAMTypeInternet()")
-
-	if branch == "" {
-		log.Info().Msg("branch not specified, using 'main' by default")
-		branch = "main"
-	}
-
-	providerUrl := fmt.Sprintf(
-		"https://raw.githubusercontent.com/Keyfactor/%s/%s/integration-manifest.json",
-		repo,
-		branch,
-	)
-	log.Debug().Str("providerUrl", providerUrl).
-		Msg("Getting PAM Type from Internet")
-	response, err := http.Get(providerUrl)
-	if err != nil {
-		log.Error().Err(err).
-			Str("providerUrl", providerUrl).
-			Msg("error getting PAM Type from Internet")
-		return nil, err
-	}
-	log.Trace().Interface("httpResponse", response).
-		Msg("GetPAMTypeInternet")
-
-	//check response status code is 200
-	if response.StatusCode != 200 {
-		return nil, fmt.Errorf("invalid response status: %s", response.Status)
-	}
-
-	defer response.Body.Close()
-
-	log.Debug().Msg("Parsing PAM response")
-	manifest, iErr := io.ReadAll(response.Body)
-	if iErr != nil {
-		log.Error().Err(iErr).
-			Str("providerUrl", providerUrl).
-			Msg("unable to read PAM response")
-		return nil, iErr
-	}
-	log.Trace().Interface("manifest", manifest).Send()
-
-	var manifestJson map[string]interface{}
-	log.Debug().Msg("Converting PAM response to JSON")
-	jErr := json.Unmarshal(manifest, &manifestJson)
-	if jErr != nil {
-		log.Error().Err(jErr).
-			Str("providerUrl", providerUrl).
-			Msg("invalid integration-manifest.json provided")
-		return nil, jErr
-	}
-	log.Debug().Msg("Parsing manifest response for PAM type config")
-	pamTypeJson := manifestJson["about"].(map[string]interface{})["pam"].(map[string]interface{})["pam_types"].(map[string]interface{})[providerName]
-	if pamTypeJson == nil {
-		// Check if only one PAM Type is defined
-		pamTypeJson = manifestJson["about"].(map[string]interface{})["pam"].(map[string]interface{})["pam_types"].(map[string]interface{})
-		if len(pamTypeJson.(map[string]interface{})) == 1 {
-			for _, v := range pamTypeJson.(map[string]interface{}) {
-				pamTypeJson = v
-			}
-		} else {
-			return nil, fmt.Errorf("unable to find PAM type %s in manifest on %s", providerName, providerUrl)
-		}
-	}
-
-	log.Trace().Interface("pamTypeJson", pamTypeJson).Send()
-	log.Debug().Msg("returning: GetPAMTypeInternet()")
-	return pamTypeJson, nil
-}
-
-func GetTypeFromInternet[T JSONImportableObject](providerName string, repo string, branch string, returnType *T) (
-	*T,
-	error,
-) {
-	log.Debug().Str("providerName", providerName).
-		Str("repo", repo).
-		Str("branch", branch).
-		Msg("entered: GetTypeFromInternet()")
-
-	log.Debug().Msg("call: GetPAMTypeInternet()")
-	manifestJSON, err := GetPAMTypeInternet(providerName, repo, branch)
-	log.Debug().Msg("returned: GetPAMTypeInternet()")
-	if err != nil {
-		log.Error().Err(err).Send()
-		return new(T), err
-	}
-
-	log.Debug().Msg("Converting PAM Type from manifest to bytes")
-	manifestJSONBytes, jErr := json.Marshal(manifestJSON)
-	if jErr != nil {
-		log.Error().Err(jErr).Send()
-		return new(T), jErr
-	}
-
-	var objectFromJSON T
-	log.Debug().Msg("Converting PAM Type from bytes to JSON")
-	mErr := json.Unmarshal(manifestJSONBytes, &objectFromJSON)
-	if mErr != nil {
-		log.Error().Err(mErr).Send()
-		return new(T), mErr
-	}
-
-	log.Debug().Msg("returning: GetTypeFromInternet()")
-	return &objectFromJSON, nil
-}
-
-func GetTypeFromConfigFile[T JSONImportableObject](filename string, returnType *T) (*T, error) {
-	log.Debug().Str("filename", filename).
-		Msg("entered: GetTypeFromConfigFile()")
-
-	log.Debug().Str("filename", filename).
-		Msg("Opening PAM Type config file")
-	file, err := os.Open(filename)
-	if err != nil {
-		log.Error().Err(err).Send()
-		return new(T), err
-	}
-
-	var rawData map[string]interface{}
-	decoder := json.NewDecoder(file)
-	dErr := decoder.Decode(&rawData)
-	if dErr != nil {
-		log.Error().Err(dErr).Send()
-		return new(T), dErr
-	}
-
-	if _, ok := rawData["about"]; ok {
-		// If the file contains the full manifest, extract the PAM type config
-		log.Debug().Msg("Parsing PAM Type config from manifest file")
-		about := rawData["about"].(map[string]interface{})
-		pam := about["pam"].(map[string]interface{})
-		pamTypes := pam["pam_types"].(map[string]interface{})
-		var pamTypeConfig interface{}
-		if len(pamTypes) == 1 {
-			for _, v := range pamTypes {
-				pamTypeConfig = v
-			}
-		} else {
-			return new(T), fmt.Errorf("multiple PAM types found in manifest file, please provide a file with a single PAM type definition")
-		}
-
-		log.Debug().Msg("Converting PAM Type config from manifest to bytes")
-		pamTypeConfigBytes, jErr := json.Marshal(pamTypeConfig)
-		if jErr != nil {
-			log.Error().Err(jErr).Send()
-			return new(T), jErr
-		}
-
-		var objectFromManifest T
-		log.Debug().Msg("Converting PAM Type config from bytes to JSON")
-		mErr := json.Unmarshal(pamTypeConfigBytes, &objectFromManifest)
-		if mErr != nil {
-			log.Error().Err(mErr).Send()
-			return new(T), mErr
-		}
-
-		log.Debug().Msg("returning: GetTypeFromConfigFile()")
-		return &objectFromManifest, nil
-	}
-
-	// Rewind file pointer to beginning
-	_, sErr := file.Seek(0, io.SeekStart)
-	if sErr != nil {
-		log.Error().Err(sErr).Send()
-		return new(T), sErr
-	}
-
-	// If the file contains only the PAM type config
-	var objectFromFile T
-	log.Debug().Msg("Decoding PAM Type config file")
-	decoder = json.NewDecoder(file)
-	dErr = decoder.Decode(&objectFromFile)
-	if dErr != nil {
-		log.Error().Err(dErr).Send()
-		return new(T), dErr
-	}
-
-	log.Debug().Msg("returning: GetTypeFromConfigFile()")
-	return &objectFromFile, nil
-}
-
 func init() {
 	var filePath string
 	var name string
@@ -918,6 +386,7 @@ func init() {
 		"Path to a JSON file containing the PAM Type Object Data.",
 	)
 	pamTypesCreateCmd.Flags().StringVarP(&name, "name", "n", "", "Name of the PAM Provider Type.")
+	pamTypesCreateCmd.Flags().BoolVarP(&all, "all", "a", false, "Create all PAM Provider Types.")
 	pamTypesCreateCmd.Flags().StringVarP(&repo, "repo", "r", "", "Keyfactor repository name of the PAM Provider Type.")
 	pamTypesCreateCmd.Flags().StringVarP(
 		&branch,
